requirements.txt
picamera2
opencv-python-headless
numpy
RPi.GPIO
